pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "voxflow"
version = "1.0.0"
description = "VoxFlow — Lokalne rozpoznawanie mowy (Speech-to-Text) dla Windows. 100% offline, wspiera polski i angielski."
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "AI Evolution Polska"},
]
keywords = ["speech-to-text", "whisper", "polish", "dictation", "offline", "voxflow"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Win32 (MS Windows)",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: Microsoft :: Windows",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Sound/Audio :: Speech",
    "Natural Language :: Polish",
    "Natural Language :: English",
]
dependencies = [
    "faster-whisper>=1.0.0,<2.0.0",
    "sounddevice>=0.4.6,<1.0.0",
    "numpy>=1.24.0,<3.0.0",
    "customtkinter>=5.2.0,<6.0.0",
    "keyboard>=0.13.5,<1.0.0",
    "pystray>=0.19.5,<1.0.0",
    "Pillow>=10.0.0,<12.0.0",
    "pyperclip>=1.8.2,<2.0.0",
    "scipy>=1.11.0,<2.0.0",
]

[project.urls]
Homepage = "https://github.com/AI-Evolution-Polska/VoxFlow"
Repository = "https://github.com/AI-Evolution-Polska/VoxFlow"
Issues = "https://github.com/AI-Evolution-Polska/VoxFlow/issues"
Changelog = "https://github.com/AI-Evolution-Polska/VoxFlow/blob/main/CHANGELOG.md"

[project.scripts]
voxflow = "voxflow.main:main"

[tool.setuptools.packages.find]
include = ["voxflow*"]

[tool.setuptools.package-data]
voxflow = ["*.py"]
